@maxxwizard 2017-07-24
I didn't realize there was a dev branch until today. That's an intense rewrite! Excited to see where you guys bring this tool to.