@bill-long 2017-07-24
OK it turns out we DO need to skip the first sample. I have some additional changes that will make fixing that either, so we will address that in an upcoming pull request.